Add better error handling, change buffer handling.
Initially I was waiting for the buffer to become empty, and
then feed one complete buffer full of data to the plotter.
This made the plotting pause momentarily while we detect
the empty buffer, and start filling again.

Now I'm querying for the actual amount of free space in the
plotter. But now it turned out that if I just send a few bytes
to top up the buffer, this may actually be too much, combined
with my next query for new data.

So, therefore I wait until the buffer (which is 1024 bytes on
my plotter) reaches 256 free bytes, and then I'll push in the
number of free bytes minus 128, so that I don't overflow it.

This worked so far, no further buffer overruns happened.